% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genesym.R, R/mousernaage.R
\name{convert_genesym}
\alias{convert_genesym}
\title{convert_genesym}
\usage{
convert_genesym()

convert_genesym()
}
\arguments{
\item{()}{함수의 input parameter}
}
\description{
convert ENSEMBL ID to Official gene symbol

convert ENSEMBL ID to Official gene symbol
}
