% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousernaage.R, R/prereq.R
\name{prereq}
\alias{prereq}
\title{prereq}
\usage{
prereq()

prereq()
}
\arguments{
\item{()}{함수의 input parameter}
}
\description{
installing pre-requisite packages

installing pre-requisite packages
}
