% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousernaage.R, R/rmdup.R
\name{rmdup}
\alias{rmdup}
\title{rmdup}
\usage{
rmdup()

rmdup()
}
\arguments{
\item{()}{함수의 input parameter}
}
\description{
Removing duplicated genes and NAs

Removing duplicated genes and NAs
}
